fix avio_alloc_context leak and move deprecated implementations
